add picolisp ebuild
It currently has QA notice about RWX sections in some binaries, and
another one about pre-stripped binaries. Since it's a super-compact lisp
interpreter, written largely in assembly, it seems to be only natural
for it to have polymorphing abilities. Hardened Gentoo wouldn't be happy
about it, I suppose.

But there is a possibility that this is just underspecified assembly
sections, and actually a bug. We need to consult picolisp developers
regarging this.